FIX backward compatibility was broken
