+++ sklearn/preprocessing/__init__.py
@@ -20,6 +20,7 @@ from .label import LabelBinarizer
 from .label import LabelEncoder
 
 from .imputation import Imputer
+from .weights import balance_weights
 
 __all__ = [
     'Binarizer',
